checkpoint: renames "Byte/exports.ts" to "Byte/Encoded/exports.ts"
