+++ src/library/Sequence/Byte/index.ts
@@ -1,1 +1,1 @@
-export * as Sequence_Byte from './exports.ts';
+export * as Sequence_Byte from './Encoded/exports.ts';
